Refactor view folder file structure
