+++ src/pages/main.ts
@@ -1,9 +1,8 @@
 import template from "./main.template";
-import ColorPicker from "../views/color-picker";
-import PrevGradient from "../views/prev-gradient";
-import ColorList from "../views/color-list";
-import ChangeOptions from "../views/change-options";
-import CodeViewer from "../views/code-viewer";
+import PrevGradient from "../views/PrevGradient/prev-gradient";
+import ColorList from "../views/ColorList/color-list";
+import ChangeOptions from "../views/ChangeOptions/change-options";
+import CodeViewer from "../views/CodeViewer/code-viewer";
 
 import { title } from "../data/site-meta.json";
 import { AnyObject } from "../types/common";
@@ -25,13 +24,11 @@ export default class MainPage {
 
   private initialize() {
     const prevGradient = new PrevGradient("#prev-gradient", this._data);
-    // const colorPicker = new ColorPicker("#color-picker", this._data);
     const colorList = new ColorList("#color-list", this._data);
     const changeOptions = new ChangeOptions("#change-option", this._data);
     const codeViewer = new CodeViewer("#code-viewer", this._data);
 
     this._fields.push(prevGradient);
-    // this._fields.push(colorPicker);
     this._fields.push(colorList);
     this._fields.push(changeOptions);
     this._fields.push(codeViewer);
