+++ src/views/ChangeOptions/index.ts
@@ -1,0 +1,3 @@
+import ChangeOptions from "./change-options";
+
+export default ChangeOptions
