+++ src/views/CodeViewer/index.ts
@@ -1,0 +1,3 @@
+import CodeViewer from "./code-viewer";
+
+export default CodeViewer
