+++ src/views/ColorList/index.ts
@@ -1,0 +1,3 @@
+import ColorList from "./color-list";
+
+export default ColorList
