+++ src/views/ColorPicker/color-picker.template.ts
