+++ src/views/ColorPicker/index.ts
@@ -1,0 +1,3 @@
+import ColorPicker from "./color-picker";
+
+export default ColorPicker
