+++ src/views/GradientBar/gradient-bar.template.ts
