+++ src/views/GradientBar/index.ts
@@ -1,0 +1,3 @@
+import GradientBar from "./gradient-bar";
+
+export default GradientBar
