+++ src/views/PrevGradient/index.ts
@@ -1,0 +1,3 @@
+import PrevGradient from "./prev-gradient";
+
+export default PrevGradient
