+++ src/views/PrevGradient/prev-gradient.template.ts
