+++ src/views/PrevGradient/prev-gradient.ts
@@ -1,6 +1,6 @@
-import CoreView from "./core-view";
+import CoreView from "../core-view";
 import template from "./prev-gradient.template";
-import { AnyObject } from "../types/common";
+import { AnyObject } from "../../types/common";
 
 class PrevGradient extends CoreView {
   private _data: AnyObject;
